+++ core-ng/src/main/java/core/framework/api/util/Properties.java
@@ -1,8 +1,10 @@
 package core.framework.api.util;
 
-import java.io.FileNotFoundException;
+import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.Reader;
 import java.io.UncheckedIOException;
 import java.util.Map;
 import java.util.Optional;
@@ -14,10 +16,11 @@ public final class Properties {
     final Map<String, String> properties = Maps.newHashMap();
 
     public void load(String path) {
-        try (InputStream stream = Thread.currentThread().getContextClassLoader().getResourceAsStream(path)) {
-            if (stream == null) throw new FileNotFoundException("property not found in classpath, classpath=" + path);
-            java.util.Properties properties = new java.util.Properties();
-            properties.load(stream);
+        InputStream stream = Thread.currentThread().getContextClassLoader().getResourceAsStream(path);
+        if (stream == null) throw Exceptions.error("can not find property in classpath, classpath={}", path);
+        java.util.Properties properties = new java.util.Properties();
+        try (Reader reader = new BufferedReader(new InputStreamReader(stream, Charsets.UTF_8))) {
+            properties.load(reader);
             properties.forEach((key, value) -> {
                 String previous = this.properties.putIfAbsent((String) key, (String) value);
                 if (previous != null)