+++ core-ng/src/main/java/core/framework/impl/template/CDNFunction.java
@@ -1,8 +1,0 @@
-package core.framework.impl.template;
-
-/**
- * @author neo
- */
-public interface CDNFunction {
-    String url(String url);
-}