+++ core-ng/src/main/java/core/framework/impl/template/MessageFunction.java
@@ -1,8 +1,0 @@
-package core.framework.impl.template;
-
-/**
- * @author neo
- */
-public interface MessageFunction {
-    String message(String key);
-}